Fixing bug in log which allows log to overflow the DDR and crash the node.
git-svn-id: http://warpproject.org/svn/WARP@3925 dd934ca3-7c2e-414f-9084-136bfdd87501

+++ ReferenceDesigns/w3_802.11/c/wlan_mac_high_framework/wlan_mac_event_log.c
@@ -690,7 +690,7 @@ int  event_log_get_next_empty_address( u32 size, u32 * address ) {
 		allocation_mutex = 1;
 
 		// Compute the end address of the newly allocated entry
-	    end_address = log_next_address + size;
+	    end_address = (u64)(log_next_address) + (u64)(size);
 
 	    // Check if the log has wrapped
 	    if ((log_next_address > log_oldest_address) ||
